Refactoring continues
