+++ server/queues/tasks/ExportHTMLZipTask.ts
@@ -1,10 +1,17 @@
+import JSZip from "jszip";
 import { FileOperationFormat } from "@shared/types";
 import { Collection } from "@server/models";
-import { archiveCollections } from "@server/utils/zip";
+import { addCollectionsToArchive } from "@server/utils/zip";
 import ExportTask from "./ExportTask";
 
 export default class ExportHTMLZipTask extends ExportTask {
   public async export(collections: Collection[]) {
-    return await archiveCollections(collections, FileOperationFormat.HTMLZip);
+    const zip = new JSZip();
+
+    return await addCollectionsToArchive(
+      zip,
+      collections,
+      FileOperationFormat.HTMLZip
+    );
   }
 }
