+++ server/queues/tasks/ExportMarkdownZipTask.ts
@@ -1,11 +1,15 @@
+import JSZip from "jszip";
 import { FileOperationFormat } from "@shared/types";
 import { Collection } from "@server/models";
-import { archiveCollections } from "@server/utils/zip";
+import { addCollectionsToArchive } from "@server/utils/zip";
 import ExportTask from "./ExportTask";
 
 export default class ExportMarkdownZipTask extends ExportTask {
   public async export(collections: Collection[]) {
-    return await archiveCollections(
+    const zip = new JSZip();
+
+    return await addCollectionsToArchive(
+      zip,
       collections,
       FileOperationFormat.MarkdownZip
     );
