+++ server/utils/zip.ts
@@ -1,5 +1,5 @@
 import path from "path";
-import JSZip, { JSZipObject } from "jszip";
+import JSZip from "jszip";
 import { FileOperationFormat } from "@shared/types";
 import Logger from "@server/logging/Logger";
 import Attachment from "@server/models/Attachment";
@@ -11,29 +11,6 @@ import ZipHelper from "./ZipHelper";
 import { serializeFilename } from "./fs";
 import parseAttachmentIds from "./parseAttachmentIds";
 import { getFileByKey } from "./s3";
-
-type ItemType = "collection" | "document" | "attachment";
-
-export type Item = {
-  path: string;
-  dir: string;
-  name: string;
-  depth: number;
-  metadata: Record<string, any>;
-  type: ItemType;
-  item: JSZipObject;
-};
-
-export type FileTreeNode = {
-  /** The title, extracted from the file name */
-  title: string;
-  /** The file name including extension */
-  name: string;
-  /** The full path to within the zip file */
-  path: string;
-  /** The nested children */
-  children: FileTreeNode[];
-};
 
 async function addDocumentTreeToArchive(
   zip: JSZip,
@@ -59,7 +36,20 @@ async function addDocumentTreeToArchive(
     });
 
     for (const attachment of attachments) {
-      await addImageToArchive(zip, attachment.key);
+      try {
+        const img = await getFileByKey(attachment.key);
+        if (img) {
+          ZipHelper.addToArchive(zip, attachment.key, img as Blob, {
+            createFolders: true,
+          });
+        }
+      } catch (err) {
+        Logger.error(
+          `Failed to add attachment to archive: ${attachment.key}`,
+          err
+        );
+      }
+
       text = text.replace(attachment.redirectUrl, encodeURI(attachment.key));
     }
 
@@ -67,7 +57,7 @@ async function addDocumentTreeToArchive(
 
     const extension = format === FileOperationFormat.HTMLZip ? "html" : "md";
 
-    title = ZipHelper.safeAddToArchive(zip, `${title}.${extension}`, text, {
+    title = ZipHelper.addToArchive(zip, `${title}.${extension}`, text, {
       date: document.updatedAt,
       comment: JSON.stringify({
         createdAt: document.createdAt,
@@ -85,33 +75,11 @@ async function addDocumentTreeToArchive(
   }
 }
 
-/**
- * Adds the content of a file in remote storage to the given zip file.
- *
- * @param zip JSZip object to add to
- * @param key path to file in S3 storage
- */
-async function addImageToArchive(zip: JSZip, key: string) {
-  try {
-    const img = await getFileByKey(key);
-
-    // @ts-expect-error Blob
-    zip.file(key, img, {
-      createFolders: true,
-    });
-  } catch (err) {
-    Logger.error("Error loading image attachment from S3", err, {
-      key,
-    });
-  }
-}
-
-export async function archiveCollections(
+export async function addCollectionsToArchive(
+  zip: JSZip,
   collections: Collection[],
   format: FileOperationFormat
 ) {
-  const zip = new JSZip();
-
   for (const collection of collections) {
     if (collection.documentStructure) {
       const folder = zip.folder(serializeFilename(collection.name));
